+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [0.2.0] 2021-04-6
+
+### New features
+
+- Update components design
+- Add new features for the components
+- Add transparent background for components
+
 ## [0.1.1] 2021-03-31
 
 ### Bug fixing